Move EditableTag to seperate file
